--- radix-engine/build.rs
@@ -1,12 +1,0 @@
-use std::process::Command;
-
-fn main() {
-    let status = Command::new("cargo")
-        .current_dir("./tests/everything")
-        .args(["build", "--target", "wasm32-unknown-unknown", "--release"])
-        .status()
-        .unwrap();
-    if !status.success() {
-        panic!()
-    }
-}
