+++ radix-engine/tests/call/src/lib.rs
@@ -1,0 +1,1 @@
+pub mod move_resource;
