--- radix-engine/tests/everything/src/lib.rs
@@ -1,8 +1,0 @@
-pub mod badge;
-pub mod bucket;
-pub mod component;
-pub mod context;
-pub mod lazy_map;
-pub mod move_resource;
-pub mod package;
-pub mod resource_def;
