+++ radix-engine/tests/lazy_map/src/lib.rs
@@ -1,0 +1,1 @@
+pub mod lazy_map;
