+++ radix-engine/tests/package/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "everything"
+name = "package"
 version = "0.1.0"
 edition = "2018"
