+++ radix-engine/tests/package/src/lib.rs
@@ -1,0 +1,1 @@
+pub mod package;
