+++ radix-engine/tests/resource_def/src/lib.rs
@@ -1,0 +1,1 @@
+pub mod resource_def;
